requirements.txt
Django==1.11.5
django-filebrowser-no-grappelli==3.7.2
django-tinymce4-lite==1.4.1
jsmin==2.2.2
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
pytz==2017.2
six==1.11.0
